Update cart item

+++ src/main/java/com/vpe/finalstore/cart/entities/Cart.java
@@ -51,11 +51,11 @@ public class Cart {
         var cartItem = getItem(variant.getVariantId());
 
         if (cartItem != null) {
-            cartItem.setQuantity(cartItem.getQuantity() + 1);
+            cartItem.setQuantity((short) (cartItem.getQuantity() + 1));
         } else {
             cartItem = new CartItem();
             cartItem.setVariant(variant);
-            cartItem.setQuantity(1);
+            cartItem.setQuantity((short) 1);
             cartItem.setCart(this);
             cartItems.add(cartItem);
         }
